Add a more thorough explanation of usage
Not sure if this is in the right place in the README, would love feedback.

I think this would be a valuable addition to the documentation because I had a bit of a hard time
figuring out what to do. A vague memory from elm-conf that elm packages are github repos
was the key insight, and then it was guess-and-check to get it working.